feat: finished signup with email verify token

+++ app/(auth)/verify/[verifyToken]/page.tsx
@@ -12,12 +12,14 @@ interface PageProps {
 export default async function Page({params} : PageProps) {
   const token = await params.verifyToken
   const redisClient = getRedisClient()
+  if(token.length <= 0) {
+    return notFound()
+  }
   const isTokenValid = (await redisClient.json.get(`signuptemp:${token}`)) as SignUpTemp | null
-
+  
   if(!isTokenValid) {
     return notFound()
   }
-  await redisClient.del(`signuptemp:${token}`)
   redisClient.destroy()
   const {email , name , password}  = isTokenValid
   await sql.query('INSERT INTO users (name, email, password , "emailVerified") VALUES ($1, $2, $3 , NOW() )', [name, email, bcrypt.hashSync(password, 10)])
